dashboard/requirements.txt
streamlit
pandas
requests
plotly
pymongo
python-dotenv
